requirements.txt
altair==4.2.2
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.33
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.38.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==1.0.0
gsheetsdb==0.1.13.1
gspread==5.7.2
gspread-dataframe==3.3.0
gurobipy==10.0.0
idna==3.4
importlib-metadata==6.0.0
ipopt==1.0.3
ipykernel==6.15.2
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyter-server==1.18.1
jupyterlab-pygments==0.1.2
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
munch==2.5.0
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.5.0
nest-asyncio==1.5.5
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
pao==1.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
Pyomo==6.4.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyUtilib==6.0.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.19.0
terminado==0.13.1
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.1.1
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
zipp==3.15.0
